requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
pydantic
python-dotenv
httpx>=0.24,<0.28
httpx>=0.24,<0.28
certifi